Serialize per-run record (config + metrics + git SHA) as run.json
New tracking.py: git_sha() / git_dirty() capture provenance; log_run()
writes a complete, self-describing run.json into the run dir (config via
asdict, test + CV metrics, SHA, dirty flag). Replaces the hardcoded, lossy
experiments.csv DictWriter -- the record is now complete by construction,
so adding a config field never touches a writer.

load_runs() rebuilds the cross-run comparison table on demand from the
run.json files (pd.json_normalize), auto-unioning keys -- so it now
includes drop_cols/keep_cols and the CV metrics the old CSV omitted.

CV results cast to float for JSON serialization. experiments.csv left as
frozen history (annotated next).

+++ src/models/evaluate.py
@@ -56,10 +56,10 @@ def run_cv(model: Pipeline,
     cv = cross_validate(estimator=model, X=X_train, y=y_train, cv=skf, scoring=scoring)
 
     return {
-        "roc_auc_mean": cv["test_roc_auc"].mean(),
-        "roc_auc_std":  cv["test_roc_auc"].std(),
-        "pr_auc_mean":  cv["test_average_precision"].mean(),
-        "pr_auc_std":   cv["test_average_precision"].std(),
+        "roc_auc_mean": float(cv["test_roc_auc"].mean()),
+        "roc_auc_std":  float(cv["test_roc_auc"].std()),
+        "pr_auc_mean":  float(cv["test_average_precision"].mean()),
+        "pr_auc_std":   float(cv["test_average_precision"].std()),
     }
     
 # Plot ROC curve
